zadanie 3.6

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -2,14 +2,15 @@ package com.kodilla.basic_assertion;
 
 public class Calculator {
 
-    public int sum(int a, int b) {
+    public double sum(double a, double b) {
         return a + b;
     }
-    public int subtract(int a, int b) {
+    public double subtract(double a, double b) {
         return a - b;
     }
-    public int squared(int a) {
+    public double squared(double a) {
         return a * a;
     }
+
 
 }
